GotOpt Parse
This patch adds the GotOpt Parser, a usability layer on top of the pure
getop functionality. The GotOpt Parser receives parsed options on a
channel, a channel which can also receive any parsing errors, and will
always send any remaining command line argument as a []string as the
last element on the channel before it is closed. Please be aware if
there are no non-option arguments to send, no string array is sent prior
to the channel being closed. See the gotopt_test.go for an example of
how to use the GotOpt Parser.